% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_itapop.R
\name{get_itapop}
\alias{get_itapop}
\title{Get and aggregate Italian resident population data}
\usage{
get_itapop(
  years,
  geo_level = c("municipality", "province", "region", "national"),
  by_sex = TRUE,
  by_citizenship = TRUE,
  by_age = TRUE,
  include_names = TRUE,
  force_download = FALSE,
  overwrite = FALSE,
  as_data_frame = TRUE
)
}
\arguments{
\item{years}{Integer vector. The year(s) of data to retrieve (e.g., \code{2019:2022}).}

\item{geo_level}{Character. The desired geographic level of detail.
Can be one of \code{"municipality"} (default), \code{"province"}, \code{"region"}, or \code{"national"}.}

\item{by_sex}{Logical. Should the population be split by sex? Default is \code{TRUE}.}

\item{by_citizenship}{Logical. Should the population be split by citizenship? Default is \code{TRUE}.}

\item{by_age}{Logical. Should the population be split by single year of age? Default is \code{TRUE}.}

\item{include_names}{Logical. If \code{TRUE}, automatically joins the aggregated data
with the official ISTAT geographic names (regions, provinces, municipalities). Default is \code{TRUE}.}

\item{force_download}{Logical. If \code{TRUE}, bypasses the interactive prompt and forces the download of missing datasets. Defaults to \code{FALSE}.}

\item{overwrite}{Logical. If \code{TRUE}, forces the redownload of the datasets even if they already exist in the cache. Defaults to \code{FALSE}.}

\item{as_data_frame}{Logical. If \code{TRUE}, returns a base R \code{data.frame} instead of a \code{data.table}. Defaults to \code{TRUE}.}
}
\value{
A \code{data.table} (or \code{data.frame} if \code{as_data_frame = TRUE}) containing the requested and aggregated population counts,
optionally enriched with geographic names.
}
\description{
This is the main high-level function of the \code{itapop} package. It automatically
downloads (or loads from cache) the requested years of data and aggregates them
according to the user's specifications.
}
\details{
By default, the function returns the most detailed data available (municipality level,
distinguishing by sex, citizenship, and single year of age). Set the demographic
arguments to \code{FALSE} to aggregate the population counts over those dimensions.
}
\examples{
\dontrun{
# 1. Get raw, fully detailed data for 2022
df_2022_raw <- get_itapop(2022)

# 2. Get national population for 2020-2022, aggregated by age and sex
df_nat <- get_itapop(
  years = 2020:2022,
  geo_level = "national",
  by_citizenship = FALSE
)

# 3. Get regional totals for 2023 with geographic names included
df_reg_names <- get_itapop(2023,
  geo_level = "region",
  by_sex = FALSE,
  by_citizenship = FALSE,
  by_age = FALSE,
  include_names = TRUE
)

# 4. Force update of cached data for 2023 and return as a base data.frame
df_2023_updated <- get_itapop(2023, overwrite = TRUE, as_data_frame = TRUE)
}
}
