% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_data.r
\docType{data}
\name{municipality_code}
\alias{municipality_code}
\title{Italian Municipalities (Comuni) Codes and Names}
\format{
A tibble with 7,869 rows and 4 variables:
\describe{
\item{municipality_code}{ISTAT municipal code (numeric).}
\item{municipality_name}{Name of the municipality (character).}
\item{province_code}{ISTAT provincial/supramunicipal code linking to \code{province_code} (numeric).}
\item{region_code}{ISTAT regional code linking to \code{region_code} (numeric).}
}
}
\source{
\url{https://www.istat.it/classificazione/codici-dei-comuni-delle-province-e-delle-regioni/}
}
\usage{
municipality_code
}
\description{
A dataset containing the official ISTAT codes and names of Italian municipalities,
along with their corresponding province and region codes. The boundaries and
codes are harmonized.
}
\keyword{datasets}
